pyproject.toml
[project]
name = "fastapi-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "tortoise-orm>=0.25.1",
]
